+++ src/main/kotlin/kr/hhplus/be/server/support/client/ConcertApiClient.kt
@@ -29,6 +29,13 @@ class ConcertApiClient(
 
     // 외부 서비스 호출 가정이므로, 트랜잭션 분리
     @Transactional(propagation = Propagation.REQUIRES_NEW)
+    fun findConcerts(ids: List<Long>): List<Concert> {
+        log.info("[외부 서비스 호출 가정] - 콘서트 목록 조회")
+        return concertService.findConcerts(ids)
+    }
+
+    // 외부 서비스 호출 가정이므로, 트랜잭션 분리
+    @Transactional(propagation = Propagation.REQUIRES_NEW)
     fun completePaymentSeat(concertSeatId: Long) {
         log.info("[외부 서비스 호출 가정] - 좌석 결제 완료 처리")
         concertService.completePaymentSeat(concertSeatId)
